+++ Frameworks/AVFoundation.framework/AVQueuePlayerInternal.h
@@ -5,7 +5,7 @@
 @class NSMutableArray;
 
 @interface AVQueuePlayerInternal : NSObject {
-    int beginModificationCount;
+    long long beginModificationCount;
     NSMutableArray *queuedModifications;
 }
