+++ Frameworks/AVFoundation.framework/AVVideoCompositionCoreAnimationToolInternal.h
@@ -8,7 +8,7 @@
     CALayer *animationLayer;
     int auxiliaryTrackID;
     NSArray *postProcessingVideoLayers;
-    int way;
+    long long way;
 }
 
 @end
