+++ Frameworks/AddressBookUI.framework/ABNumberFormatting.h
@@ -5,7 +5,7 @@
 @interface ABNumberFormatting : NSObject {
 }
 
-+ (id)localizedStringWithInt:(int)arg1;
-+ (id)newLocalizedStringWithInt:(int)arg1;
++ (id)localizedStringWithInt:(long long)arg1;
++ (id)newLocalizedStringWithInt:(long long)arg1;
 
 @end
