+++ Frameworks/AddressBookUI.framework/ABPrivacyAdditions.h
@@ -5,6 +5,6 @@
 @interface ABPrivacyAdditions : NSObject {
 }
 
-+ (void)updateAccessDeniedView:(id)arg1 forAuthorizationStatus:(long)arg2;
++ (void)updateAccessDeniedView:(id)arg1 forAuthorizationStatus:(long long)arg2;
 
 @end
