+++ Frameworks/AddressBookUI.framework/ABUIDataCollectorLogger.h
@@ -5,8 +5,8 @@
 @interface ABUIDataCollectorLogger : NSObject {
 }
 
-- (void)addValue:(int)arg1 forScalarKey:(id)arg2;
+- (void)addValue:(long long)arg1 forScalarKey:(id)arg2;
 - (void)setValue:(double)arg1 forDistributionKey:(id)arg2;
-- (void)setValue:(int)arg1 forScalarKey:(id)arg2;
+- (void)setValue:(long long)arg1 forScalarKey:(id)arg2;
 
 @end
