+++ Frameworks/CoreFoundation.framework/__NSFastEnumerationEnumerator.h
@@ -5,8 +5,8 @@
 @class <NSFastEnumeration>;
 
 @interface __NSFastEnumerationEnumerator : NSEnumerator {
-    unsigned int _count;
-    unsigned int _mut;
+    unsigned long long _count;
+    unsigned long long _mut;
     <NSFastEnumeration> *_obj;
     id _origObj;
 }
