+++ Frameworks/CoreFoundation.framework/__NSPlaceholderDate.h
@@ -13,7 +13,7 @@
 - (id)initWithTimeIntervalSinceReferenceDate:(double)arg1;
 - (oneway void)release;
 - (id)retain;
-- (unsigned int)retainCount;
+- (unsigned long long)retainCount;
 - (double)timeIntervalSinceReferenceDate;
 
 @end
