+++ Frameworks/EventKitUI.framework/EKEventEditableDetailGroup.h
@@ -5,6 +5,6 @@
 @interface EKEventEditableDetailGroup : EKEventDetailGroup {
 }
 
-- (id)itemAtIndex:(unsigned int)arg1;
+- (id)itemAtIndex:(unsigned long long)arg1;
 
 @end
