+++ Frameworks/Foundation.framework/NSCorrectionCheckingResult.h
@@ -13,8 +13,8 @@
 - (id)description;
 - (void)encodeWithCoder:(id)arg1;
 - (id)initWithCoder:(id)arg1;
-- (id)initWithRange:(struct _NSRange { unsigned int x1; unsigned int x2; })arg1 replacementString:(id)arg2 alternativeStrings:(id)arg3;
-- (id)resultByAdjustingRangesWithOffset:(int)arg1;
+- (id)initWithRange:(struct _NSRange { unsigned long long x1; unsigned long long x2; })arg1 replacementString:(id)arg2 alternativeStrings:(id)arg3;
+- (id)resultByAdjustingRangesWithOffset:(long long)arg1;
 - (unsigned long long)resultType;
 
 @end
