+++ Frameworks/Foundation.framework/NSKeyValueFastMutableSet1.h
@@ -8,11 +8,11 @@
     NSKeyValueNonmutatingSetMethodSet *_nonmutatingMethods;
 }
 
-+ (struct { unsigned int x1; id x2[4]; }*)_proxyNonGCPoolPointer;
++ (struct { unsigned long long x1; id x2[4]; }*)_proxyNonGCPoolPointer;
 
 - (id)_proxyInitWithContainer:(id)arg1 getter:(id)arg2;
 - (void)_proxyNonGCFinalize;
-- (unsigned int)count;
+- (unsigned long long)count;
 - (id)member:(id)arg1;
 - (id)objectEnumerator;
