+++ Frameworks/MediaPlayer.framework/MPAVErrorResolver.h
@@ -13,7 +13,7 @@
 - (void).cxx_destruct;
 - (id)delegate;
 - (void)resolveError:(id)arg1;
-- (void)sendDidResolveError:(id)arg1 withResolution:(int)arg2;
+- (void)sendDidResolveError:(id)arg1 withResolution:(long long)arg2;
 - (void)setDelegate:(id)arg1;
 
 @end
