+++ Frameworks/MediaPlayer.framework/MPAVSystemRoutingController.h
@@ -5,7 +5,7 @@
 @class NSObject<OS_dispatch_queue>, NSPointerArray;
 
 @interface MPAVSystemRoutingController : NSObject {
-    int _clientDiscoveryMode;
+    long long _clientDiscoveryMode;
     NSPointerArray *_clientRoutingControllers;
     int _debugNotifyToken;
     NSObject<OS_dispatch_queue> *_serialQueue;
