+++ Frameworks/MediaPlayer.framework/MPMovieErrorLog.h
@@ -10,13 +10,13 @@
 
 @property(readonly) NSArray * events;
 @property(readonly) NSData * extendedLogData;
-@property(readonly) unsigned int extendedLogDataStringEncoding;
+@property(readonly) unsigned long long extendedLogDataStringEncoding;
 
 - (void).cxx_destruct;
 - (id)_initWithAVItemErrorLog:(id)arg1;
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
 - (id)events;
 - (id)extendedLogData;
-- (unsigned int)extendedLogDataStringEncoding;
+- (unsigned long long)extendedLogDataStringEncoding;
 
 @end
