+++ Frameworks/MediaPlayer.framework/MPMovieErrorLogEvent.h
@@ -12,7 +12,7 @@
 @property(readonly) NSDate * date;
 @property(readonly) NSString * errorComment;
 @property(readonly) NSString * errorDomain;
-@property(readonly) int errorStatusCode;
+@property(readonly) long long errorStatusCode;
 @property(readonly) NSString * playbackSessionID;
 @property(readonly) NSString * serverAddress;
 
@@ -23,7 +23,7 @@
 - (id)date;
 - (id)errorComment;
 - (id)errorDomain;
-- (int)errorStatusCode;
+- (long long)errorStatusCode;
 - (id)playbackSessionID;
 - (id)serverAddress;
