+++ Frameworks/MessageUI.framework/MFAddressConcatenator.h
@@ -22,9 +22,9 @@
 
 - (id)andNMoreFormat;
 - (id)andNMoreNoEllipsisFormat;
-- (id)commaSeparatedAddressListWithAddressCount:(unsigned int)arg1 prefixForAddressAtIndex:(id)arg2 stringForAddressAtIndex:(id)arg3 lengthValidationBlock:(id)arg4;
+- (id)commaSeparatedAddressListWithAddressCount:(unsigned long long)arg1 prefixForAddressAtIndex:(id)arg2 stringForAddressAtIndex:(id)arg3 lengthValidationBlock:(id)arg4;
 - (void)dealloc;
-- (void)getCommaSeparatedAddressList:(out id*)arg1 andListSuffix:(out id*)arg2 withAddressCount:(unsigned int)arg3 prefixForAddressAtIndex:(id)arg4 stringForAddressAtIndex:(id)arg5 lengthValidationBlock:(id)arg6;
+- (void)getCommaSeparatedAddressList:(out id*)arg1 andListSuffix:(out id*)arg2 withAddressCount:(unsigned long long)arg3 prefixForAddressAtIndex:(id)arg4 stringForAddressAtIndex:(id)arg5 lengthValidationBlock:(id)arg6;
 - (id)nAddressesFormat;
 - (id)oneAddressFormat;
 - (void)setAndNMoreFormat:(id)arg1;
