+++ Frameworks/MessageUI.framework/_MFAttachmentScalingDataConsumer.h
@@ -12,7 +12,7 @@
 @property MFConditionLock * lock;
 @property MFGenericAttachmentStore * store;
 
-- (int)appendData:(id)arg1;
+- (long long)appendData:(id)arg1;
 - (id)lock;
 - (void)setLock:(id)arg1;
 - (void)setStore:(id)arg1;
