+++ Frameworks/MessageUI.framework/_MFCountableMatchesContext.h
@@ -5,18 +5,18 @@
 @class NSMutableSet;
 
 @interface _MFCountableMatchesContext : NSObject {
-    int _maxCount;
+    long long _maxCount;
     struct __CFDictionary { } *_peopleCount;
     NSMutableSet *_popularPeople;
 }
 
-@property int maxCount;
+@property long long maxCount;
 
 - (void)countInstances:(id)arg1 usingPredicate:(id)arg2;
 - (void)dealloc;
 - (id)highestMatches;
 - (id)init;
-- (int)maxCount;
-- (void)setMaxCount:(int)arg1;
+- (long long)maxCount;
+- (void)setMaxCount:(long long)arg1;
 
 @end
